Remove flow-network related models
