--- pkg/v1/model/producer.go
@@ -1,46 +1,0 @@
-package model
-
-var HistoryTypeMap = map[HistoryType]int8{
-	HistoryTypeCov:            0,
-	HistoryTypeInterval:       0,
-	HistoryTypeCovAndInterval: 0,
-}
-
-var HistoryTypeCovMap = map[HistoryType]int8{
-	HistoryTypeCov:            0,
-	HistoryTypeCovAndInterval: 0,
-}
-
-/*
-Producer a producer is a placeholder to register an object to enable consumers to a producer for example is a point,
-something that makes data, and the subscriber would have a consumer to it, like grafana reading and writing to it
-from edge to cloud or wires over rest(peer to peer)
-*/
-type Producer struct {
-	CommonUUID
-	Name string `json:"name" gorm:"uniqueIndex:idx_producers_name_stream_uuid"`
-	CommonDescription
-	CommonEnable
-	CommonSyncUUID
-	ProducerThingName       string         `json:"producer_thing_name,omitempty"`  // e.g. point.name, user will understand what name it is
-	ProducerThingUUID       string         `json:"producer_thing_uuid,omitempty"`  // e.g. point.uuid
-	ProducerThingClass      string         `json:"producer_thing_class,omitempty"` // e.g. point.thing_class, i.e. point, job etc.
-	ProducerThingType       string         `json:"producer_thing_type,omitempty"`  // e.g. point.thing_type, i.e. temp, rssi, voltage etc.
-	ProducerApplication     string         `json:"producer_application,omitempty"`
-	CommonCurrentWriterUUID                // if the point for example is read only the writer.uuid would be the point.uuid, i.e.: itself, so in this case there is no writer or writer clone
-	StreamUUID              string         `json:"stream_uuid,omitempty" gorm:"type:string references streams;not null;uniqueIndex:idx_producers_name_stream_uuid"`
-	WriterClones            []*WriterClone `json:"writer_clones,omitempty" gorm:"constraint:OnDelete:CASCADE"`
-	Tags                    []*Tag         `json:"tags,omitempty" gorm:"many2many:producers_tags;constraint:OnDelete:CASCADE"`
-	CommonCreated
-}
-
-// ProducerBody could be a local network, job or alarm and so on
-type ProducerBody struct {
-	CommonThingClass // point, job
-	CommonThingRef
-	CommonThingType             // for example temp, rssi, voltage
-	FlowNetworkUUID string      `json:"flow_network_uuid,omitempty"`
-	ProducerUUID    string      `json:"producer_uuid,omitempty"`
-	StreamUUID      string      `json:"stream_uuid,omitempty"`
-	Payload         interface{} `json:"payload"`
-}
